pyproject.toml
[project]
name = "mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
]
